requirements.txt
python-dotenv==0.12.0
ConfigArgParse==1.2.3
aioconsole==0.3.1
aiofiles==0.6.0
requests==2.22.0
async-timeout==3.0.1
contextvars==2.4
more-itertools==8.6.0
anyio==2.0.2
